+++ src/consumer/component/package-json.js
@@ -17,9 +17,10 @@ import GeneralError from '../../error/general-error';
 import JSONFile from './sources/json-file';
 import npmRegistryName from '../../utils/bit/npm-registry-name';
 import componentIdToPackageName from '../../utils/bit/component-id-to-package-name';
+import DataToPersist from './sources/data-to-persist';
 
 // the instance comes from bit-javascript PackageJson class
-export type PackageJsonInstance = { write: Function, bit?: Object };
+export type PackageJsonInstance = { write: Function, bit?: Object, componentRootFolder: string };
 
 /**
  * Add components as dependencies to root package.json
@@ -92,7 +93,7 @@ function getPackageDependencyValue(
   return convertToValidPathForPackageManager(rootDirRelative);
 }
 
-async function getPackageDependency(consumer: Consumer, dependencyId: BitId, parentId: BitId): Promise<?string> {
+function getPackageDependency(consumer: Consumer, dependencyId: BitId, parentId: BitId): ?string {
   const parentComponentMap = consumer.bitMap.getComponent(parentId);
   const dependencyComponentMap = consumer.bitMap.getComponentIfExist(dependencyId);
   return getPackageDependencyValue(dependencyId, parentComponentMap, dependencyComponentMap);
@@ -155,7 +156,7 @@ async function write(
   writeBitDependencies?: boolean = false,
   excludeRegistryPrefix?: boolean
 ): Promise<PackageJsonInstance> {
-  const packageJson: PackageJsonInstance = await preparePackageJsonToWrite(
+  const { packageJson, distPackageJson } = preparePackageJsonToWrite(
     consumer,
     component,
     bitDir,
@@ -164,31 +165,32 @@ async function write(
     excludeRegistryPrefix
   );
   await packageJson.write({ override });
+  if (distPackageJson) await distPackageJson.write({ override });
   return packageJson;
 }
 
-async function preparePackageJsonToWrite(
+function preparePackageJsonToWrite(
   consumer: Consumer,
   component: Component,
   bitDir: string,
   override?: boolean = true,
   writeBitDependencies?: boolean = false,
   excludeRegistryPrefix?: boolean
-): Promise<PackageJsonInstance> {
+): { packageJson: PackageJsonInstance, distPackageJson: ?PackageJsonInstance } {
   logger.debug(`package-json.write. bitDir ${bitDir}. override ${override.toString()}`);
   const PackageJson = consumer.driver.getDriver(false).PackageJson;
-  const getBitDependencies = async (dependencies: Dependencies) => {
+  const getBitDependencies = (dependencies: Dependencies) => {
     if (!writeBitDependencies) return {};
     const dependenciesPackages = dependencies.get().map(async (dep) => {
-      const packageDependency = await getPackageDependency(consumer, dep.id, component.id);
+      const packageDependency = getPackageDependency(consumer, dep.id, component.id);
       return [dep.id.toStringWithoutVersion(), packageDependency];
     });
-    return R.fromPairs(await Promise.all(dependenciesPackages));
+    return R.fromPairs(dependenciesPackages);
   };
-  const bitDependencies = await getBitDependencies(component.dependencies);
-  const bitDevDependencies = await getBitDependencies(component.devDependencies);
-  const bitCompilerDependencies = await getBitDependencies(component.compilerDependencies);
-  const bitTesterDependencies = await getBitDependencies(component.testerDependencies);
+  const bitDependencies = getBitDependencies(component.dependencies);
+  const bitDevDependencies = getBitDependencies(component.devDependencies);
+  const bitCompilerDependencies = getBitDependencies(component.compilerDependencies);
+  const bitTesterDependencies = getBitDependencies(component.testerDependencies);
   const registryPrefix = component.bindingPrefix || npmRegistryName();
   const name = excludeRegistryPrefix
     ? componentIdToPackageName(component.id, component.bindingPrefix, false)
@@ -219,14 +221,14 @@ async function preparePackageJsonToWrite(
   };
   const packageJson = getPackageJsonInstance(bitDir);
 
+  let distPackageJson;
   if (!component.dists.isEmpty() && !component.dists.areDistsInsideComponentDir) {
     const distRootDir = component.dists.distsRootDir;
     if (!distRootDir) throw new GeneralError('component.dists.distsRootDir is not defined yet');
-    const distPackageJson = getPackageJsonInstance(distRootDir);
-    await distPackageJson.write({ override });
+    distPackageJson = getPackageJsonInstance(distRootDir);
   }
 
-  return packageJson;
+  return { packageJson, distPackageJson };
 }
 
 async function updateAttribute(
@@ -322,6 +324,16 @@ async function getPackageJsonObject(dir: string): Promise<?Object> {
   }
 }
 
+function addPackageJsonDataToPersist(packageJson: PackageJsonInstance, dataToPersist: DataToPersist) {
+  const packageJsonPath = composePath(packageJson.componentRootFolder);
+  const jsonFile = JSONFile.load({
+    base: packageJson.componentRootFolder,
+    path: packageJsonPath,
+    content: packageJson
+  });
+  dataToPersist.addFile(jsonFile);
+}
+
 async function writePackageJsonFromObject(dir: string, data: Object) {
   return fs.outputJSON(composePath(dir), data, { spaces: 2 });
 }
@@ -336,6 +348,7 @@ export {
   changeDependenciesToRelativeSyntax,
   write,
   preparePackageJsonToWrite,
+  addPackageJsonDataToPersist,
   addComponentsWithVersionToRoot,
   updateAttribute,
   addWorkspacesToPackageJson,