+++ src/links/node-modules-linker.js
@@ -19,6 +19,7 @@ import Symlink from './symlink';
 import DataToPersist from '../consumer/component/sources/data-to-persist';
 import LinkFile from './link-file';
 import ComponentsList from '../consumer/component/components-list';
+import { preparePackageJsonToWrite, addPackageJsonDataToPersist } from '../consumer/component/package-json';
 
 type LinkDetail = { from: string, to: string };
 export type LinksResult = {
@@ -168,7 +169,7 @@ export default class NodeModuleLinker {
         this.dataToPersist.addSymlink(Symlink.makeInstance(file, dest, componentId));
       }
     });
-    this._populateLinkToMainFile(component);
+    this._createPackageJsonForAuthor(component);
   }
   /**
    * When the dists is outside the components directory, it doesn't have access to the node_modules of the component's
@@ -281,28 +282,15 @@ export default class NodeModuleLinker {
   }
 
   /**
-   * Link from node_modules/@bit/component-name/index.js to the component's main file.
+   * create package.json on node_modules/@bit/component-name/package.json with a property 'main'
+   * pointing to the component's main file.
    * It is needed for Authored components only.
    * Since an authored component doesn't have rootDir, it's impossible to symlink to the component directory.
    * It makes it easier for Author to use absolute syntax between their own components.
    */
-  _populateLinkToMainFile(component: Component) {
-    component.dists.updateDistsPerConsumerBitJson(component.id, this.consumer, component.componentMap);
-    const mainFile = component.dists.calculateMainDistFileForAuthored(component.mainFile, this.consumer);
-    const indexFileName = getIndexFileName(mainFile);
-    const dest = path.join(getNodeModulesPathOfComponent(component.bindingPrefix, component.id), indexFileName);
-    const destRelative = this._getPathRelativeRegardlessCWD(path.dirname(dest), mainFile);
-    const fileContent = getLinkToFileContent(destRelative);
-    if (fileContent) {
-      // otherwise, the file type is not supported, no need to write anything
-      const linkFile = LinkFile.load({
-        filePath: dest,
-        content: fileContent,
-        srcPath: mainFile,
-        componentId: component.id,
-        override: true
-      });
-      this.dataToPersist.addFile(linkFile);
-    }
+  _createPackageJsonForAuthor(component: Component) {
+    const dest = path.join(getNodeModulesPathOfComponent(component.bindingPrefix, component.id));
+    const { packageJson } = preparePackageJsonToWrite(this.consumer, component, dest, true);
+    addPackageJsonDataToPersist(packageJson, this.dataToPersist);
   }
 }